feat: replan and implement the code construction within the CodeEditor module, clarifying the functions of each internal module and their interrelationships
