+++ spx-gui/src/components/editor/code-editor/ai-chat.ts
