+++ spx-gui/src/components/editor/code-editor/compiler.ts
