+++ spx-gui/src/components/editor/code-editor/ui/ToolItem.vue
