+++ spx-gui/src/components/editor/code-editor/ui/code-text-editor/monaco.ts
